requirements.txt
Flask~=1.1.1
slackeventsapi~=2.1.0
slackclient~=2.2.0
